feat: Implement keyboard-driven screenshot tool
This commit provides a simple, focused Chrome extension to capture a region of the screen using a keyboard shortcut. This replaces the previous, more complex UI-based implementation based on user feedback.

The extension's workflow is as follows:
1. The user presses a keyboard shortcut (default: Ctrl+Shift+S or Cmd+Shift+S).
2. The cursor changes to a crosshair, allowing the user to click and drag to select an area of the page.
3. Upon releasing the mouse, the selected region is captured.
4. The resulting screenshot is immediately opened in a new tab for viewing and saving.

This implementation uses Manifest V3 and consists of a background script to handle commands and capture, and a content script to handle the on-page selection UI.
