--- .gitignore
@@ -1,14 +1,0 @@
-# Dependencies
-/node_modules
-
-# Build artifacts
-/dist
-/build
-
-# Logs
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# Vite cache
-.vite/
