--- package.json
@@ -1,30 +1,0 @@
-{
-  "name": "liquid-glass-screenshots",
-  "version": "1.0.0",
-  "description": "Chrome extension for customizable screenshots.",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build && cp manifest.json dist/"
-  },
-  "keywords": [
-    "chrome-extension",
-    "react",
-    "mui"
-  ],
-  "author": "Jules",
-  "license": "UNLICENSED",
-  "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.18",
-    "@mui/material": "^5.14.18",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-react": "^4.2.0",
-    "vite": "^5.0.0"
-  }
-}
