--- public/popup.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Liquid Glass Screenshots</title>
-    <!-- The bundler will inject the necessary styles -->
-  </head>
-  <body>
-    <div id="root"></div>
-    <script src="/src/popup/popup.jsx" type="module"></script>
-  </body>
-</html>
