--- src/content/content.css
@@ -1,24 +1,0 @@
-/* This class is added to the body to enable the crosshair cursor. */
-body.liquid-glass-selecting {
-  cursor: crosshair !important;
-}
-
-/* Styles for the selection box drawn by the user. */
-#liquid-glass-selection-box {
-  position: fixed;
-  border: 2px dashed #00aaff;
-  background-color: rgba(0, 170, 255, 0.2);
-  z-index: 99999999; /* Ensure it's on top of everything */
-  box-sizing: border-box;
-}
-
-/* A simple overlay to dim the page slightly during selection. */
-#liquid-glass-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(0, 0, 0, 0.1);
-  z-index: 99999998; /* Just below the selection box */
-}
