--- src/popup/popup.css
@@ -1,26 +1,0 @@
-/* Set a background on the root to be blurred */
-html {
-  background-image: url('https://source.unsplash.com/random/800x600?gradient');
-  background-size: cover;
-}
-
-/* Make the body transparent so the html background shows through */
-body {
-  background-color: transparent;
-  width: 380px;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
-  color: white;
-}
-
-/* Style scrollbars to be more subtle */
-::-webkit-scrollbar {
-  width: 8px;
-  height: 8px;
-}
-::-webkit-scrollbar-thumb {
-  background: rgba(255, 255, 255, 0.4);
-  border-radius: 4px;
-}
-::-webkit-scrollbar-track {
-  background: rgba(0, 0, 0, 0.1);
-}
